+++ src/core/utils/index.ts
@@ -1,1 +1,0 @@
-export { default as getClosestAnchorParent } from './getClosestAnchorParent.js';